# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Tadpole OS API
  version: 1.1.58
  description: >
    API surface generated from server-rs/src/router.rs.
servers:
  - url: http://127.0.0.1:8000
    description: Default local Rust engine
security:
  - bearerAuth: []
...
paths:
  /v1/agents:
    get:
      tags: [agents]
      summary: "Get agents"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponseList'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    post:
      tags: [agents]
      summary: "Create agent"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EngineAgent'
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
  /v1/agents/{id}:
    delete:
      tags: [agents]
      summary: "Delete agent"
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    put:
      tags: [agents]
      summary: "Update agent"
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentConfigUpdate'
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccessResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    GenericResponse:
      type: object
      additionalProperties: true
    ProblemDetails:
      type: object
      required: [type, title, status, detail, severity]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, nullable: true }
        error_code: { type: string, nullable: true }
        help_link: { type: string, nullable: true }
        severity: { type: string }
    AgentResponse:
      type: object
      required: [id, name, role, department, status, model, provider, budgetUsd, costUsd, isHealthy, isBankrupt, skills, version]
      properties:
        id: { type: string }
        name: { type: string }
        role: { type: string }
        department: { type: string }
        status: { type: string }
        model: { type: string }
        provider: { type: string }
        budgetUsd: { type: number }
        costUsd: { type: number }
        isHealthy: { type: boolean }
        isBankrupt: { type: boolean }
        skills:
          type: array
          items: { type: string }
        createdAt: { type: string, format: date-time, nullable: true }
        version: { type: integer }
    AgentResponseList:
      type: object
      required: [data, page, perPage, total, totalPages, _links]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/AgentResponse' }
        page: { type: integer }
        perPage: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }
        _links:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/HateoasLink'
    HateoasLink:
      type: object
      required: [href]
      properties:
        href: { type: string }
        method: { type: string, nullable: true }
    EngineAgent:
      type: object
      required: [identity, models, economics, health, capabilities, state]
      properties:
        identity: { $ref: '#/components/schemas/AgentIdentity' }
        models: { $ref: '#/components/schemas/AgentModels' }
        economics: { $ref: '#/components/schemas/AgentEconomics' }
        health: { $ref: '#/components/schemas/AgentHealth' }
        capabilities: { $ref: '#/components/schemas/AgentCapabilities' }
        state: { $ref: '#/components/schemas/AgentState' }
        metadata: { type: object, additionalProperties: true }
        createdAt: { type: string, format: date-time, nullable: true }
        requiresOversight: { type: boolean }
        voiceId: { type: string, nullable: true }
        voiceEngine: { type: string, nullable: true }
        sttEngine: { type: string, nullable: true }
        connectorConfigs:
          type: array
          items: { type: object }
        version: { type: integer }
    AgentIdentity:
      type: object
      required: [id, name, role, department, description, category]
      properties:
        id: { type: string }
        name: { type: string }
        role: { type: string }
        department: { type: string }
        description: { type: string }
        category: { type: string }
        themeColor: { type: string, nullable: true }
    AgentEconomics:
      type: object
      required: [budgetUsd, costUsd, tokensUsed, tokenUsage]
      properties:
        budgetUsd: { type: number }
        costUsd: { type: number }
        tokensUsed: { type: integer }
        tokenUsage: { $ref: '#/components/schemas/TokenUsage' }
    TokenUsage:
      type: object
      required: [input, output, total]
      properties:
        input: { type: integer }
        output: { type: integer }
        total: { type: integer }
    AgentHealth:
      type: object
      required: [status, failureCount]
      properties:
        status: { type: string }
        failureCount: { type: integer }
        lastFailureAt: { type: string, format: date-time, nullable: true }
        heartbeatAt: { type: string, format: date-time, nullable: true }
    AgentModels:
      type: object
      required: [model]
      properties:
        modelId: { type: string, nullable: true }
        model: { $ref: '#/components/schemas/ModelConfig' }
        model2: { type: string, nullable: true }
        model3: { type: string, nullable: true }
        modelConfig2: { $ref: '#/components/schemas/ModelConfig', nullable: true }
        modelConfig3: { $ref: '#/components/schemas/ModelConfig', nullable: true }
        activeModelSlot: { type: integer, nullable: true }
    ModelConfig:
      type: object
      required: [provider, modelId]
      properties:
        provider: { type: string }
        modelId: { type: string }
        apiKey: { type: string, nullable: true }
        temperature: { type: number, nullable: true }
        baseUrl: { type: string, nullable: true }
        reasoningDepth: { type: integer, nullable: true }
        actThreshold: { type: number, nullable: true }
    AgentCapabilities:
      type: object
      required: [skills, workflows, mcpTools]
      properties:
        skills:
          type: array
          items: { type: string }
        workflows:
          type: array
          items: { type: string }
        mcpTools:
          type: array
          items: { type: string }
        skillManifest: { type: object, nullable: true }
    AgentState:
      type: object
      required: [workingMemory, currentReasoningTurn]
      properties:
        activeMission: { type: object, nullable: true }
        currentTask: { type: string, nullable: true }
        workingMemory: { type: object }
        currentReasoningTurn: { type: integer }
    AgentConfigUpdate:
      type: object
      properties:
        name: { type: string }
        role: { type: string }
        department: { type: string }
        provider: { type: string }
        modelId: { type: string }
        modelConfig: { $ref: '#/components/schemas/ModelConfig' }
        model2: { type: string }
        model3: { type: string }
        apiKey: { type: string }
        systemPrompt: { type: string }
        temperature: { type: number }
        baseUrl: { type: string }
        reasoningDepth: { type: integer }
        actThreshold: { type: number }
        themeColor: { type: string }
        budgetUsd: { type: number }
        externalId: { type: string }
        skills:
          type: array
          items: { type: string }
        workflows:
          type: array
          items: { type: string }
        mcpTools:
          type: array
          items: { type: string }
        activeModelSlot: { type: integer }
        modelConfig2: { $ref: '#/components/schemas/ModelConfig' }
        modelConfig3: { $ref: '#/components/schemas/ModelConfig' }
        voiceId: { type: string }
        voiceEngine: { type: string }
        sttEngine: { type: string }
        category: { type: string }
        requiresOversight: { type: boolean }
        connectorConfigs:
          type: array
          items: { type: object }
        metadata: { type: object, additionalProperties: true }
...
    GenericSuccessResponse:
      type: object
      required: [status]
      properties:
        status: { type: string }
...